+++ bootique-jetty/src/main/java/com/nhl/bootique/jetty/MappedFilter.java
@@ -1,5 +1,7 @@
 package com.nhl.bootique.jetty;
 
+import java.util.Collections;
+import java.util.Map;
 import java.util.Set;
 
 import javax.servlet.Filter;
@@ -7,11 +9,8 @@ import javax.servlet.Filter;
 /**
  * @since 0.11
  */
-public class MappedFilter {
+public class MappedFilter extends MappedWebArtifact<Filter> {
 
-	private Filter filter;
-	private Set<String> urlPatterns;
-	private String name;
 	private int order;
 
 	public MappedFilter(Filter filter, Set<String> urlPatterns, int order) {
@@ -32,18 +31,31 @@ public class MappedFilter {
 	 *            ordering will be an outer filter and will be called first.
 	 */
 	public MappedFilter(Filter filter, Set<String> urlPatterns, String name, int order) {
-		this.filter = filter;
-		this.urlPatterns = urlPatterns;
+		this(filter, urlPatterns, name, Collections.emptyMap(), order);
+	}
+
+	/**
+	 * @since 0.13
+	 * @param filter
+	 *            a filter to install inside of Jetty.
+	 * @param urlPatterns
+	 *            URL patterns that this filter will respond to.
+	 * @param name
+	 *            filter name. If null, Jetty will assign its own name.
+	 * @param params
+	 *            filter init parameters map.
+	 * @param order
+	 *            an order of the filter among all the filters in a given app.
+	 *            If two filters match the same request, filter with lower
+	 *            ordering will be an outer filter and will be called first.
+	 */
+	public MappedFilter(Filter filter, Set<String> urlPatterns, String name, Map<String, String> params, int order) {
+		super(filter, urlPatterns, name, params);
 		this.order = order;
-		this.name = name;
 	}
 
 	public Filter getFilter() {
-		return filter;
-	}
-
-	public Set<String> getUrlPatterns() {
-		return urlPatterns;
+		return getArtifact();
 	}
 
 	/**
@@ -56,11 +68,4 @@ public class MappedFilter {
 		return order;
 	}
 
-	/**
-	 * @since 0.13
-	 * @return an optional filter name.
-	 */
-	public String getName() {
-		return name;
-	}
 }