+++ bootique-jetty/src/main/java/com/nhl/bootique/jetty/server/MappedServletFactory.java
@@ -2,10 +2,13 @@ package com.nhl.bootique.jetty.server;
 
 import static java.util.Arrays.asList;
 
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
 import javax.servlet.Servlet;
+import javax.servlet.annotation.WebInitParam;
 import javax.servlet.annotation.WebServlet;
 
 import com.nhl.bootique.jetty.MappedServlet;
@@ -29,7 +32,15 @@ public class MappedServletFactory {
 
 		String name = wsAnnotation.name() != null && wsAnnotation.name().length() > 0 ? wsAnnotation.name() : null;
 		Set<String> urlPatterns = new HashSet<>(asList(wsAnnotation.urlPatterns()));
-		return new MappedServlet(servlet, urlPatterns, name);
+
+		Map<String, String> initParams = new HashMap<>();
+
+		WebInitParam[] paramsArray = wsAnnotation.initParams();
+		if (paramsArray != null) {
+			asList(paramsArray).forEach(p -> initParams.put(p.name(), p.value()));
+		}
+
+		return new MappedServlet(servlet, urlPatterns, name, initParams);
 	}
 
 }